backend/requirement.txt
Flask 
Flask-CORS 
python-dotenv 
google-generativeai 
langchain 
langchain-google-genai